chore(aws-amplify): rename adapterCore to adapter-core (#13149)
