+++ packages/aws-amplify/__tests__/adapterCore/authProvidersFactories/cognito/createAWSCredentialsAndIdentityIdProvider.test.ts
@@ -10,7 +10,7 @@ import {
 	AuthConfig,
 	KeyValueStorageInterface,
 } from '@aws-amplify/core';
-import { createAWSCredentialsAndIdentityIdProvider } from '../../../../src/adapterCore';
+import { createAWSCredentialsAndIdentityIdProvider } from '../../../../src/adapter-core';
 
 jest.mock('@aws-amplify/auth/cognito');
