+++ packages/aws-amplify/__tests__/adapterCore/authProvidersFactories/cognito/createUserPoolsTokenProvider.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@aws-amplify/auth/cognito';
 
 import { AuthConfig, KeyValueStorageInterface } from '@aws-amplify/core';
-import { createUserPoolsTokenProvider } from '../../../../src/adapterCore';
+import { createUserPoolsTokenProvider } from '../../../../src/adapter-core';
 
 jest.mock('@aws-amplify/auth/cognito');
