+++ packages/aws-amplify/__tests__/adapterCore/runWithAmplifyServerContext.test.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { runWithAmplifyServerContext } from '../../src/adapterCore';
+import { runWithAmplifyServerContext } from '../../src/adapter-core';
 import {
 	createAmplifyServerContext,
 	destroyAmplifyServerContext,
