+++ packages/aws-amplify/adapter-core/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "aws-amplify/adapter-core",
-	"types": "../dist/esm/adapterCore/index.d.ts",
-	"main": "../dist/cjs/adapterCore/index.js",
-	"module": "../dist/esm/adapterCore/index.mjs",
+	"types": "../dist/esm/adapter-core/index.d.ts",
+	"main": "../dist/cjs/adapter-core/index.js",
+	"module": "../dist/esm/adapter-core/index.mjs",
 	"sideEffects": false
 }
