+++ packages/aws-amplify/package.json
@@ -151,9 +151,9 @@
 			"react-native": "./src/push-notifications/pinpoint/index.ts"
 		},
 		"./adapter-core": {
-			"types": "./dist/esm/adapterCore/index.d.ts",
-			"import": "./dist/esm/adapterCore/index.mjs",
-			"require": "./dist/cjs/adapterCore/index.js"
+			"types": "./dist/esm/adapter-core/index.d.ts",
+			"import": "./dist/esm/adapter-core/index.mjs",
+			"require": "./dist/cjs/adapter-core/index.js"
 		},
 		"./package.json": "./package.json"
 	},
@@ -226,7 +226,7 @@
 				"./dist/esm/push-notifications/pinpoint/index.d.ts"
 			],
 			"adapter-core": [
-				"./dist/esm/adapterCore/index.d.ts"
+				"./dist/esm/adapter-core/index.d.ts"
 			]
 		}
 	},
