+++ packages/aws-amplify/src/adapter-core/authProvidersFactories/cognito/createAWSCredentialsAndIdentityIdProvider.ts
