+++ packages/aws-amplify/src/adapter-core/authProvidersFactories/cognito/createUserPoolsTokenProvider.ts
