+++ packages/aws-amplify/src/adapter-core/authProvidersFactories/cognito/index.ts
