+++ packages/aws-amplify/src/adapter-core/index.ts
