+++ packages/aws-amplify/src/adapter-core/storageFactories/index.ts
